+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 1.2
 
 - Added InventoryManager.java and ItemCreator.java
+- Added ``getBlocksBetweenLocations(Location location1, Location location2)``, ``getSmallerLocation(Location location1, Location location2)``, ``getBiggerLocation(Location location1, Location location2)``, and ``isLocationUnderBlocks(Location location)``
 
 # Version 1.1